todoapi/requirements/base.txt
Flask-Cors~=1.10.3
Flask-Migrate~=3.1.0
Flask-SQLAlchemy~=2.5.1
Flask~=1.1.2
python-dateutil~=2.8.1
pytz~=2020.4
waitress~=1.4.4